audiofilters, audiodelays: fix silence fill, clipping bound and loop counters
memset(word_buffer, 32768, ...) writes only the low byte of its value, and
32768 & 0xff is 0. So the 'fill with silence' path filled the buffer with
zeros -- which for unsigned samples is full negative deflection, not
silence. It writes 0x8000 words in a loop now.

Distortion clamped with MIN(MAX(word, -32767), 32768). The upper bound
becomes -32768 on the cast to int16_t below it, so a positive overdrive
came out as a full negative spike.

The biquad filter chain stores objs_len as a size_t but counted over it
with uint8_t in four places: the reset, tick and process helpers in
audiofilters/__init__.c and the per-block filter loop in Filter.c. Past 256
Biquad objects -- 128 in the reset helper, which counts objs_len *
channel_count -- the counter wrapped back to zero and the loop never
terminated.

Not compile-checked: CIRCUITPY_AUDIOEFFECTS is off for this board, so
neither module is built here and no compiler has seen these edits.

+++ shared-module/audiofilters/Filter.c
@@ -225,7 +225,9 @@ audioio_get_buffer_result_t audiofilters_filter_get_buffer(audiofilters_filter_o
                     }
 
                     // Process biquad filters
-                    for (uint8_t j = 0; j < self->filter.objs_len; j++) {
+                    // CIRCUITPY-CHANGE: objs_len is a size_t, so a uint8_t counter wrapped
+                    // back to zero at 256 filters and the loop never terminated.
+                    for (size_t j = 0; j < self->filter.objs_len; j++) {
                         mp_obj_t filter_obj = self->filter.objs[j];
                         common_hal_synthio_biquad_tick(filter_obj);
                         for (uint8_t k = 0; k < self->base.channel_count; k++) {
